Break out the levelTwo class to its own file, plus fix bug that prevents game from resetting due to assignment to imported variable
